- Rest Api:  * postMapping set up for saving teacher schedule of a week
 - DayFinder interface made for flexible so that it also allows usage of
 the DayOfWeek

+++ src/main/java/makeitwork/mijninzet/model/TeacherSchedule/CohortWeek.java
@@ -15,7 +15,7 @@ public class CohortWeek {
     private long weekId;
 
     @Column(name = "weekNumber")
-    private int weekNumber;
+    private long weekNumber;
 
     @ManyToOne(cascade = CascadeType.PERSIST)
     @JoinColumn(name = "cohortId")
@@ -33,11 +33,11 @@ public class CohortWeek {
         this.weekId = weekId;
     }
 
-    public int getWeekNumber() {
+    public long getWeekNumber() {
         return weekNumber;
     }
 
-    public void setWeekNumber(int weekNumber) {
+    public void setWeekNumber(long weekNumber) {
         this.weekNumber = weekNumber;
     }
